Added Random NPC Dialogue + Stats
Added stats variables as well as the 8 NPC dialogues. The dialogues get randomly chosen when you interact with an NPC with the "e" key. The dialogues are also tied to the intelligence stat, and the higher your intelligence is the more you understand (Intel of <=5  is 0% understanding, Intel of >=10 is 100% understanding).